Extract Command "meta" code into its own module
As I'm about to make some important changes to the Command module,
its so much better to extract all "meta" code from the `__using__` macro
to an external module.
